requirements.txt
Flask==3.0.3
Panda
flask_cors
